+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Mask money in `filters.js`
+- Mask square area in `filters.js`
+- Mask decimal in `filters.js`
 
 ### Fixed